Add long and test case 1

+++ src/org/experis/shop/Smartphone.java
@@ -4,11 +4,11 @@ import java.math.BigDecimal;
 
 public class Smartphone extends Prodotto {
     //Attributi
-    public int imei;
+    public long imei;
     public int memory;
 
 
-    public Smartphone(String name, String brand, BigDecimal price, BigDecimal vat,int imei,int memory) {
+    public Smartphone(String name, String brand, BigDecimal price, BigDecimal vat,long imei,int memory) {
         super(name, brand, price, vat);
         this.imei = imei;
         this.memory = memory;
